Remove forced alignment code.
The move from to gcc 4.8 to gcc 4.9 for arm32 introduced a bug in this
code. The original code is trying to out smart the compiler by arch, but
we got caught. Running benchmarks, the amount of time we save by doing this
is in the nanosecond range, so just let the compiler figure things out on
it's own.

It turns out for aarch64, x86, x86_64, two of the functions produce exactly
the same code. For swapLongs, x86/x86_64 produces slightly different code
but is about the same performance.

For arm32, letting the compiler optimize also leads to about the same
performance.

Adding unit tests and benchmark code for these.

Bug: 19692084

New unit tests and benchmark code is removed because Android 5.1 source tree
is not compatible with these and just applying libcore_io_Memory.cpp patch
is enough.

Change-Id: Icbea0e59d3d52ddd628d559478a5cc903a16c7b3
Signed-off-by: arter97 <qkrwngud825@gmail.com>